Added roles dialog base class, read NTP from DHCP
